Move new thin_client code into original files
