+++ tests/multinode.rs
@@ -15,7 +15,6 @@ use solana::streamer::default_window;
 use solana::thin_client::ThinClient;
 use std::io;
 use std::io::sink;
-use std::net::UdpSocket;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, RwLock};
 use std::thread::sleep;
@@ -144,17 +143,14 @@ fn test_multi_node() {
 }
 
 fn mk_client(leader: &ReplicatedData) -> ThinClient {
-    let requests_socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-    requests_socket
-        .set_read_timeout(Some(Duration::new(1, 0)))
-        .unwrap();
-    let transactions_socket = UdpSocket::bind("0.0.0.0:0").unwrap();
+    let requests_addr = "0.0.0.0:0".parse().unwrap();
+    let transactions_addr = "0.0.0.0:0".parse().unwrap();
 
     ThinClient::new(
         leader.requests_addr,
-        requests_socket,
+        requests_addr,
         leader.transactions_addr,
-        transactions_socket,
+        transactions_addr,
     )
 }
